Release Candidate v0.7.5
- Added native support for the XFL
- Added drop down to Config UI for `league_id`
- Improved checks for `league_id` in YAML
- Code cleanup
- Improved unit tests

Thanks to @raman325 for the changes in this release.

Note:  The current version (v0.7.5) allows the Configuration Check to pass with an invalid `league_id` in YAML and will create an error in the log.  These errors must be addressed in this version as the next version (v0.7.6) will prevent the Configuration Check from passing with an invalid `league_id`.