v0.14.1 Release Candidate (#165)
- Added programmatic control of the sensor with the service `call_api` (beta)
- This enables the dynamic retrieval of information for different teams without the need to create multiple sensors.  It will simply reuse the same sensor.  This is expected to be of value to meet the needs of dynamic voice requests.
- Accepts four input parameters:  `sport_path`, `league_path`, `team_id`, and `conference_id` (optional)
- Service will dynamically change the sensor to the input parameters and refresh the sensor
- Sensor will revert to original configuration on reboot

NOTE:  This does not allow `teamtracker` to support multiple teams at the same time.  It simply allows the supported team to be changed programatically.

NOTE:  The functionality and input parameters of the `call_api` service is subject to change as new uses emerge.  If you have ideas or suggestions for specific use cases, open an Issue so they can be tracked.

Thanks to @dinki for the suggestion